omc: Safety checkpoint

+++ openm/test_templates/test_range.cpp
@@ -43,10 +43,10 @@ namespace mm {
         assert(y1 == -3);
 
         // test of to_index
-        size_t n1 = y1.to_index();
+        int n1 = y1.to_index();
         assert(n1 == 0);
 
-        size_t n2 = YEAR::to_index(0);
+        int n2 = YEAR::to_index(0);
         assert(n2 == 3);
 
         int n3 = YEAR::to_value(5);
